finally finished

+++ app/src/main/java/com/example/billsparser/BillView.kt
@@ -1,3 +1,3 @@
 package com.example.billsparser
 
-data class BillView(val debtorName: String, val debtAmount: Float)
+data class BillView(val debtorName: String, val debtAmount: Double)
